data_reader/requirements.txt
pyserial==3.5
questdb==2.0.3
